requirements.txt
colorama
pyfiglet
gspread
google-auth
